chain_test/rag_chain/src/rag_chain/gcp_integration/secrets.py
from google.cloud import secretmanager
from rag_chain.core.config import settings

def load_secret(name):
    client = secretmanager.SecretManagerServiceClient()
    name = f"projects/{settings.gcp_project_id}/secrets/{name}/versions/latest"
    response = client.access_secret_version(request={"name": name})
    return response.payload.data.decode("UTF-8")

